Use trussed::virt instead of own platform
The trussed::virt module now provides a virtual platform so we no longer
have to setup the Trussed platform manually.
